Fix doc error: genotype_prompt will not recognize "exit" line
The documentation had said to include an "exit" line at the
end of list of regions to genotype, but the genotype_prompt
function's loop does not recognize this command as ending
the file, but tries to parse it as a region.

Getting rid of this results in the program ending after there
is no more input (which is idiomatic for most programs, anyway).